% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diriv.R
\name{diriv}
\alias{diriv}
\title{Derivative of the link function}
\usage{
diriv(family, muhat)
}
\arguments{
\item{family}{A description of the link function used in the model:
\code{"gaussian"} or \code{"binomial"}}

\item{muhat}{fitted values}
}
\value{
derivative of the link function for the fitted values
}
\description{
Computes the derivative of the link function according to
the distribution family specified in the \code{"family"} parameter.
}
\author{
Ines Ortega-Fernandez, Marta Sestelo.
}
\keyword{internal}
