% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R, R/train.R
\docType{data}
\name{homoscedastic_uniform_gaussian_train}
\alias{homoscedastic_uniform_gaussian_train}
\title{Homoscedastic Uniform Gaussian synthetic test data.}
\format{
\code{homoscedastic_uniform_gaussian_train} is a data frame with 7
variables (columns).
The first three columns of the data frame contains the covariates X0, X1, X2.
The next three columns of the data frame contains the corresponding f(X) for
each covariate
The last column of the data frame contains the response variable.
contains the response variable

\code{homoscedastic_uniform_gaussian_train} is a data frame with 7
variables (columns).
The first three columns of the data frame contains the covariates X0, X1, X2.
The next three columns of the data frame contains the corresponding f(X) for
each covariate
The last column of the data frame contains the response variable.
contains the response variable
}
\usage{
homoscedastic_uniform_gaussian_train

homoscedastic_uniform_gaussian_train
}
\description{
Homoscedastic Uniform Gaussian synthetic data frame with 7 variables
and 5000 measurements for evaluation. This dataset was generated to build a
simulation scenario to evaluate NeuralGAM.

Homoscedastic Uniform Gaussian synthetic data frame with 7 variables
and 24500 measurements for training. This dataset was generated to build a
simulation scenario to evaluate NeuralGAM.
}
\details{
We considered the predictor:

{\eta = \alpha + \sum^3_{j=1} f_j(X_j)},
with
{f_1(X_1) = X^2_1}
{f_2(X_2) = 2X_2}
{f_3(X_3) = \sin(X_3)}

{$\alpha = 2}, and covariates {X_1, X_2, X_3} drawn from an uniform
distribution {U\left[-2.5, 2.5\right]} and response variable
{Y = \eta + \varepsilon} where {\varepsilon} is an homoscedastic error term
distributed in accordance to a {N(0,\sigma(x)}

We considered the predictor:

{\eta = \alpha + \sum^3_{j=1} f_j(X_j)},
with
{f_1(X_1) = X^2_1}
{f_2(X_2) = 2X_2}
{f_3(X_3) = \sin(X_3)}

{$\alpha = 2}, and covariates {X_1, X_2, X_3} drawn from an uniform
distribution {U\left[-2.5, 2.5\right]} and response variable
{Y = \eta + \varepsilon} where {\varepsilon} is an homoscedastic error term
distributed in accordance to a {N(0,\sigma(x)}
}
\examples{
library(NeuralGAM)
data(test)
head(test)

X_test = test[c('X0','X1','X2')]
fs_test = test[c('f(X0)','f(X1)','f(X2)')]
y_test = test['y']
library(NeuralGAM)
data(train)
head(train)

X_train = train[c('X0','X1','X2')]
fs_train = train[c('f(X0)','f(X1)','f(X2)')]
y_train = train['y']
}
