% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralGAM-package.R
\docType{package}
\name{neuralGAM-package}
\alias{neuralGAM-package}
\alias{_PACKAGE}
\title{neuralGAM: Interpretable Neural Network Based on Generalized Additive Models}
\description{
Neural network framework based on Generalized Additive Models from Hastie & Tibshirani (1990, ISBN:9780412343902), which trains a different neural network to estimate the contribution of each feature to the response variable. The networks are trained independently leveraging the local scoring and backfitting algorithms to ensure that the Generalized Additive Model converges and it is additive. The resultant Neural Network is a highly accurate and interpretable deep learning model, which can be used for high-risk AI practices where decision-making should be based on accountable and interpretable algorithms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://inesortega.github.io/neuralGAM/}
  \item Report bugs at \url{https://github.com/inesortega/neuralGAM/issues}
}

}
\author{
\strong{Maintainer}: Ines Ortega-Fernandez \email{iortega@gradiant.org} (\href{https://orcid.org/0000-0002-8041-6860}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Marta Sestelo \email{sestelo@uvigo.es} (\href{https://orcid.org/0000-0003-4284-6509}{ORCID}) [copyright holder]
}

}
\keyword{internal}
