% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\docType{data}
\name{train}
\alias{train}
\title{Homoscedastic Uniform Gaussian synthetic training data.}
\format{
\code{train} is a data frame with 7
variables (columns).
The first three columns of the data frame contains the covariates \eqn{X_1, X_2, X_3}
The next three columns of the data frame contains the corresponding \eqn{f(X)} for
each covariate
The last column of the data frame contains the response variable \eqn{y}
}
\usage{
data(train)
}
\description{
Homoscedastic Uniform Gaussian synthetic data frame with 7 variables
and 24500 measurements for training. This dataset was generated to build a
simulation scenario to evaluate NeuralGAM.
}
\details{
We considered the predictor: \eqn{\eta = \alpha + \sum^3_{j=1} f_j(X_j)}, with

\deqn{
f_j(X_j) =
\begin{cases}
X_j^2 \hspace{1.4cm} \text{ if } j=1 \\
2X_j \hspace{1.3cm} \text{ if } j=2 \\
\sin{X_j} \hspace{1cm} \text{ if } j=3,
\end{cases}
}
\eqn{\alpha = 2}, and covariates \eqn{X_1, X_2, X_3} drawn from an uniform
distribution \eqn{U\left[-2.5, 2.5\right]} and response variable
\eqn{Y = \eta + \varepsilon} where \eqn{\varepsilon} is an homoscedastic error term
distributed in accordance to a \eqn{N(0,\sigma(x))}
}
\examples{
library(NeuralGAM)
data(train)
head(train)

X_train <- train[c('X0','X1','X2')]
fs_train <- train[c('f(X0)','f(X1)','f(X2)')]
y_train <- train$y
}
