% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{weight}
\alias{weight}
\title{Weights}
\usage{
weight(w, muhat, family)
}
\arguments{
\item{w}{weights}

\item{muhat}{fitted values}

\item{family}{A description of the link function used in the model:
\code{"gaussian"} or \code{"binomial"}}
}
\value{
computed weights for the Local Scoring algorithm
according to the \code{"family"} distribution
}
\description{
Computes the weights for the Local Scoring Algorithm.
}
\examples{
eta <- weight(w, muhat, "gaussian")
eta <- weight(w, muhat, "binomial")
}
\author{
Ines Ortega-Fernandez, Marta Sestelo and Nora M. Villanueva.
}
