% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_feature_NN.R
\name{build_feature_NN}
\alias{build_feature_NN}
\title{Build and compile a neural network feature model}
\usage{
build_feature_NN(
  num_units,
  learning_rate = 0.001,
  activation = "relu",
  kernel_initializer = "glorot_normal",
  kernel_regularizer = NULL,
  bias_regularizer = NULL,
  bias_initializer = "zeros",
  activity_regularizer = NULL,
  loss = "mean_squared_error",
  name = NULL,
  ...
)
}
\arguments{
\item{num_units}{Defines the architecture of each neural network.
If a scalar value is provided, a single hidden layer with that number of units is used.
If a vector of values is provided, each element defines the number of units in a hidden layer.}

\item{learning_rate}{Learning rate for the neural network optimizer.}

\item{activation}{Activation function for the hidden layers. Defaults to `"relu"`.}

\item{kernel_initializer}{Kernel initializer for the Dense layers.
Defaults to Xavier Initializer (`"glorot_normal"`).}

\item{kernel_regularizer}{Optional regularizer function applied to the kernel weights.}

\item{bias_regularizer}{Optional regularizer function applied to the bias.}

\item{bias_initializer}{Optional initializer for the bias vector.}

\item{activity_regularizer}{Optional regularizer function applied to the output of the layer.}

\item{loss}{Loss function to use during training. Defaults to `"mse"`.}

\item{name}{Character string. Name assigned to the neural network corresponding to the smooth term.}

\item{...}{Additional parameters passed to `keras::optimizer_adam()`.}
}
\value{
A compiled `keras_model` object ready for training.
}
\description{
Builds and compiles a `keras` neural network for a single smooth term in a `neuralGAM` model.
The architecture (number of units, layers, activation, regularization, etc.) is configured
through the parameters inherited from `neuralGAM`.
}
\references{
Kingma, D. P., & Ba, J. (2014). Adam: A method for stochastic optimization.
arXiv preprint arXiv:1412.6980.
}
\author{
Ines Ortega-Fernandez, Marta Sestelo.
}
\keyword{internal}
