% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{deviance}
\alias{deviance}
\title{Deviance of the model}
\usage{
deviance(fit, y, family)
}
\arguments{
\item{fit}{current estimation of the response variable}

\item{y}{response variable}

\item{family}{A description of the link function used in the model: (\code{"gaussian"}) or (\code{"binomial"})}
}
\value{
the deviance of the model
}
\description{
Computes the deviance of the model according to the distribution
family specified in the \code{"family"} parameter (\code{"gaussian"})
or (\code{"binomial"})
}
\examples{
dev <- deviance(muhat, y_train, "gaussian")
dev <- deviance(muhat, y_train, "binomial")
}
