% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-dropout.R
\name{.mc_dropout_forward}
\alias{.mc_dropout_forward}
\title{Internal helper functions to compute MC dropout samples}
\usage{
.mc_dropout_forward(model, x, passes, output_dim)
}
\description{
Internal helper functions to compute MC dropout samples
}
\keyword{internal}
