% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{get_formula_elements}
\alias{get_formula_elements}
\title{Extract formula elements}
\usage{
get_formula_elements(formula)
}
\arguments{
\item{formula}{A formula object}
}
\value{
A list with the following elements:
\itemize{
\item y: The response variable
\item terms: A character vector with all model terms
\item np_terms: A character vector with non-parametric terms
\item p_terms: A character vector with parametric terms
\item np_formula: The formula for the non-parametric terms
\item p_formula: The formula for the parametric terms
\item formula: The original formula object
}
}
\description{
This function separates the model terms of a given formula into response,
all_terms, non-parametric terms and parametric terms.
}
\keyword{internal}
