% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_neuralGAM}
\alias{install_neuralGAM}
\title{Install `neuralGAM` python requirements}
\usage{
install_neuralGAM(force = FALSE)
}
\arguments{
\item{force}{Whether to force the installation of miniconda and dependencies. Deafults to FALSE}
}
\description{
Creates a conda environment (installing miniconda if required) and set ups the
Python requirements to run `neuralGAM` (Tensorflow and Keras).
}
\examples{
\dontrun{
library(neuralGAM)
install_neuralGAM()
}
}
