% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_neuralGAM}
\alias{install_neuralGAM}
\title{Install neuralGAM python requirements}
\usage{
install_neuralGAM(envname = "r-tensorflow")
}
\arguments{
\item{envname}{name of the environment to be used}
}
\description{
Install neuralGAM python requirements
}
