% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_neuralGAM}
\alias{install_neuralGAM}
\title{Install neuralGAM python requirements}
\usage{
install_neuralGAM()
}
\description{
Creates a conda environment (installing miniconda if required) and set ups the
Python requirements to run neuralGAM (Tensorflow and Keras).

Miniconda and related environments are generated in the user's cache directory
given by:

\code{tools::R_user_dir('neuralGAM', 'cache')}
}
\author{
Ines Ortega-Fernandez, Marta Sestelo
}
