% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.NeuralGAM.R
\name{plot.NeuralGAM}
\alias{plot.NeuralGAM}
\title{Visualization of \code{NeuralGAM} object}
\usage{
\method{plot}{NeuralGAM}(x = object, select = NULL, xlab = NULL, ylab = NULL, title = NULL, ...)
}
\arguments{
\item{x}{a fitted \code{NeuralGAM} object as produced by \code{NeuralGAM()}.}

\item{select}{allows to plot a set of selected terms. e.g. if you just want to plot the first term,
select="X0"}

\item{xlab}{if supplied, this value will be used as the \code{x} label for all plots}

\item{ylab}{if supplied, this value will be used as the \code{y} label for all plots}

\item{\ldots}{other graphics parameters to pass on to plotting commands.
See details for ggplot2::geom_line options}
}
\value{
Returns the partial effects plot.
}
\description{
Visualization of \code{NeuralGAM} object. Plots the learned partial effects by the NeuralGAM object.
}
\examples{
#' library(NeuralGAM)
data(train)

ngam <- NeuralGAM( y ~ X1 + s(X0) + s(X2), data = train,
num_units = 1024, family = "gaussian",
learning_rate = 0.001, bf_threshold = 0.001,
max_iter_backfitting = 10, max_iter_ls = 10
)

plot(ngam)

# Plot only a given term

plot(ngam, select="X1")

}
\author{
Ines Ortega-Fernandez, Marta Sestelo.
}
