% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.NeuralGAM.R
\name{plot.NeuralGAM}
\alias{plot.NeuralGAM}
\title{Visualization of \code{NeuralGAM} object}
\usage{
\method{plot}{NeuralGAM}(x = object, y = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{\code{NeuralGAM} object.}

\item{y}{NULL}

\item{ylab}{NULL}

\item{\ldots}{Other options.}
}
\value{
Returns the partial effects plot.
}
\description{
Visualization of \code{NeuralGAM} object. Plots the learned partial effects by the NeuralGAM object.
}
\author{
Ines Ortega-Fernandez, Marta Sestelo.
}
