% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.NeuralGAM.R
\name{plot.NeuralGAM}
\alias{plot.NeuralGAM}
\title{Visualization of \code{NeuralGAM} object}
\usage{
\method{plot}{NeuralGAM}(x = object, y = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{\code{NeuralGAM} object.}

\item{y}{NULL}

\item{ylab}{NULL}

\item{\ldots}{Other options.}
}
\value{
Returns the partial effects plot.
}
\description{
This function plots the learned partial effects by the NeuralGAM object.
}
\examples{
library(NeuralGAM)
data(train)
head(train)
X_train <- train[c('X0','X1','X2')]
y_train <- train$y

ngam <- fit_NeuralGAM(num_units = 1024, learning_rate = 0.001, x=X_train,
              y = y_train, family = "gaussian", bf_threshold=0.00001,
              ls_threshold = 0.1, max_iter_backfitting = 10,
              max_iter_ls=10)

plot(ngam)
}
\author{
Ines Ortega-Fernandez, Marta Sestelo and Nora M. Villanueva.
}
