% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.NeuralGAM}
\alias{predict.NeuralGAM}
\title{predict}
\usage{
\method{predict}{NeuralGAM}(ngam, x, family = "gaussian", ...)
}
\arguments{
\item{ngam}{a fitted `NeuralGAM` object}

\item{x}{A data frame or list containing the values of covariates at which
predictions are required.}

\item{family}{A description of the link function used in the model:
\code{"gaussian"} or \code{"binomial"}}

\item{\ldots}{Other options.}
}
\value{
model predictions
}
\description{
Takes a fitted \code{NeuralGAM} object produced by
\code{fit_NeuralGAM()} and produces predictions given a new set of values.
}
