% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.NeuralGAM.R
\name{predict.NeuralGAM}
\alias{predict.NeuralGAM}
\title{Produces predictions from a fitted \code{NeuralGAM} object}
\usage{
\method{predict}{NeuralGAM}(object, x = NULL, type = "link", terms = NULL, ...)
}
\arguments{
\item{object}{a fitted `NeuralGAM` object}

\item{x}{A data frame or list containing the values of covariates at which
predictions are required. If not provided, the function returns the predictions
for the original training data.}

\item{type}{when \code{type="link"} (default), the linear
predictor is returned. When \code{type="terms"} each component of the linear
predictor is returned separately on each column of a \code{data.frame}. When
\code{type="response"} predictions on the scale of the response are returned.}

\item{terms}{If \code{type="terms"}, then only results for the terms named
in this list will be returned. If \code{NULL} then no terms are excluded (default).}

\item{\ldots}{Other options.}
}
\value{
Predicted values according to \code{type} parameter.
}
\description{
Takes a fitted \code{NeuralGAM} object produced by
\code{NeuralGAM()} and produces predictions given a new set of values for the model covariates.
}
\examples{
# Obtain linear predictor
predict(object = ngam, x = x_test, type = "link")
# Obtain each component of the linear predictor separately on each column of a data.frame
predict(object = ngam, x = x_test, type = "terms")
# Obtain only some terms
predict(object = ngam, x = x_test, type = "terms", terms=c("X0", "X1"))
# Obtain predictions on the scale of the response
predict(object = ngam, x = x_test, type = "response")
}
