% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.NeuralGAM.R
\name{print.NeuralGAM}
\alias{print.NeuralGAM}
\title{Short \code{NeuralGAM} summary}
\usage{
\method{print}{NeuralGAM}(x = object, ...)
}
\arguments{
\item{x}{\code{NeuralGAM} object.}

\item{\ldots}{Other options.}
}
\value{
The printed output of the object
}
\description{
Default print statement for a NeuralGAM object. Prints
the distribution family, model formula, intercept value, number of terms
(neural networks), as well as neural network architecture.
}
\author{
Ines Ortega-Fernandez, Marta Sestelo.
}
