% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.NeuralGAM.R
\name{summary.NeuralGAM}
\alias{summary.NeuralGAM}
\title{\code{NeuralGAM} summary}
\usage{
\method{summary}{NeuralGAM}(object, ...)
}
\arguments{
\item{object}{\code{NeuralGAM} object.}

\item{\ldots}{Other options.}
}
\value{
The summary of the object
}
\description{
Summary of a fitted NeuralGAM object. Prints
the distribution family, model formula, intercept value, sample size,
as well as neural network architecture and training history.
}
\author{
Ines Ortega-Fernandez, Marta Sestelo.
}
